twoByteToInt fixed

+++ client/src/main/java/bgu/spl/net/impl/tftp/Util.java
@@ -27,7 +27,8 @@ public class Util {
 
     public static int twoByteToInt(byte [] bytes){
         //return ( int ) ((( int ) bytes [0]) << 8 | ( int ) ( bytes [1]) );
-        return (int) (bytes[0]) << 8 | (short) (bytes[1]) & 0x00ff;
+        return (short) (((short) bytes[0]) << 8 | (short) (bytes[1]) & 0x00ff);
+        //return (int) (bytes[0]) << 8 | (short) (bytes[1]) & 0x00ff;
 
     }
